+++ oak_restricted_kernel/src/syscall/channel.rs
@@ -19,7 +19,7 @@ use core::ffi::{c_int, c_size_t, c_ssize_t, c_void};
 use alloc::{boxed::Box, slice};
 use oak_channel::Channel;
 use oak_core::sync::OnceCell;
-use oak_restricted_kernel_runtime::syscall::Errno;
+use oak_restricted_kernel_interface::Errno;
 use spinning_top::Spinlock;
 
 trait FileDescriptor {
