+++ oak_restricted_kernel_api/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "oak_restricted_kernel_runtime"
+name = "oak_restricted_kernel_api"
 version = "0.1.0"
 authors = ["Andri Saar <andrisaar@google.com>"]
 edition = "2021"
@@ -8,6 +8,7 @@ license = "Apache-2.0"
 [dependencies]
 anyhow = { version = "*", default-features = false }
 oak_channel = { path = "../oak_channel" }
+oak_restricted_kernel_interface = { path = "../oak_restricted_kernel_interface" }
 strum = { version = "*", default-features = false, features = ["derive"] }
 
 [dev-dependencies]
