+++ oak_restricted_kernel_api/src/lib.rs
@@ -23,5 +23,4 @@ mod channel;
 mod raw_syscall;
 pub mod syscall;
 
-pub use channel::SyscallChannel;
-pub use raw_syscall::Syscall;
+pub use channel::FileDescriptorChannel;
