+++ oak_restricted_kernel_interface/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "oak_restricted_kernel_interface"
+version = "0.1.0"
+authors = ["Andri Saar <andrisaar@google.com>"]
+edition = "2021"
+license = "Apache-2.0"
+
+[dependencies]
+strum = { version = "*", default-features = false, features = ["derive"] }
